Add spidev
With this program you can test the spi-device on all boards. The
delivery parameter is the device which you want to test. Before you can
test the spi you have to connect the MISO and MOSI Pins of the device
which you want to test. For more information see the description in the
file.